+++ src/components/dialog/dialog.ts
@@ -73,7 +73,7 @@ export class CuiDialogHandler extends CuiHandler<CuiDialogArgs> implements ICuiC
 
     }
 
-    async open(): Promise<boolean> {
+    async open(args?: any): Promise<boolean> {
         if (this.checkLockAndWarn('open')) {
             return false;
         }
@@ -84,35 +84,38 @@ export class CuiDialogHandler extends CuiHandler<CuiDialogArgs> implements ICuiC
         this.isLocked = true;
         this._log.debug(`Dialog ${this.cuid}`, 'open')
         let action = this.getAction(DIALOG_OPEN_ANIMATION_CLASS);
-        return this.performAction(action, this.#timeout, this.mutate.bind(this, this.onOpen));
+        return this.performAction(action, this.#timeout, this.onOpen.bind(this, args), () => {
+            this.element.classList.add(this.activeClassName);
+            document.body.classList.add(this.#bodyClass);
+            AriaAttributes.setAria(this.element, 'aria-expanded', 'true')
+        });
     }
 
-    async close(): Promise<boolean> {
+    async close(args: any): Promise<boolean> {
         if (this.checkLockAndWarn("close") || !this.isActive()) {
             return false;
         }
         this.isLocked = true;
         this._log.debug(`Dialog ${this.cuid}`, 'close')
         let action = this.getAction(DIALOG_CLOSE_ANIMATION_CLASS);
-        return this.performAction(action, this.#timeout, this.mutate.bind(this, this.onClose));
+        return this.performAction(action, this.#timeout, this.onClose.bind(this, args), () => {
+            this.element.classList.remove(this.activeClassName);
+            document.body.classList.remove(this.#bodyClass);
+            AriaAttributes.setAria(this.element, 'aria-expanded', 'false');
+        });
     }
 
-    onClose() {
-        this.element.classList.remove(this.activeClassName);
-        document.body.classList.remove(this.#bodyClass);
-        AriaAttributes.setAria(this.element, 'aria-expanded', 'false')
+    onClose(state: any) {
         this.detachEvent(EVENTS.ON_KEYDOWN);
         this.detachEvent(EVENTS.ON_WINDOW_CLICK);
         this.emitEvent(EVENTS.ON_CLOSE, {
-            timestamp: Date.now()
+            timestamp: Date.now(),
+            state: state
         })
         this.isLocked = false;
     }
 
-    onOpen() {
-        this.element.classList.add(this.activeClassName);
-        document.body.classList.add(this.#bodyClass);
-        AriaAttributes.setAria(this.element, 'aria-expanded', 'true')
+    onOpen(state?: any) {
         if (this.args.escClose) {
             this.onEvent(EVENTS.ON_KEYDOWN, this.onEscClose.bind(this))
         }
@@ -120,23 +123,24 @@ export class CuiDialogHandler extends CuiHandler<CuiDialogArgs> implements ICuiC
             this.onEvent(EVENTS.ON_WINDOW_CLICK, this.onWindowClick.bind(this));
         }
         this.emitEvent(EVENTS.ON_OPEN, {
-            timestamp: Date.now()
+            timestamp: Date.now(),
+            state: state
         })
         this.isLocked = false;
     }
 
-    onEscClose(ev: KeyDownEvent) {
+    async onEscClose(ev: KeyDownEvent) {
         if (ev.event.key === "Escape") {
-            this.close();
+            await this.close('Clesed by key');
         }
     }
 
     onWindowClick(ev: MouseEvent) {
-        this._log.debug("Dialog windows")
         let container = this.element.querySelector(replacePrefix(CONTAINER, this.#prefix));
-        console.log(container);
         if (!container.contains((ev.target as Node))) {
-            this.close();
+            this.close('out').then(() => {
+                this._log.debug("Closed by click outside")
+            });
         }
     }