+++ src/components/scroll/scroll.ts
@@ -1,9 +1,9 @@
-import { ICuiComponent, ICuiComponentHandler, CuiObservables } from "../../core/models/interfaces";
+import { ICuiComponent, ICuiComponentHandler, CuiObservables, ICuiParsable } from "../../core/models/interfaces";
 import { CuiUtils } from "../../core/models/utils";
 import { IconBuilder } from "../icon/icon";
-import { CuiHandlerBase } from "../../app/handlers/base";
+import { CuiHandlerBase, CuiHandler } from "../../app/handlers/base";
 import { ICONS, EVENTS } from "../../core/utils/statics";
-import { is, parseAttributeString, getOffsetTop } from "../../core/utils/functions";
+import { is, parseAttributeString, getOffsetTop, getStringOrDefault } from "../../core/utils/functions";
 
 export class CuiScrollComponent implements ICuiComponent {
     attribute: string;
@@ -26,30 +26,44 @@ export interface CuiScrollAttribute {
     behavior?: 'auto' | 'smooth';
 }
 
-export class CuiScrollHandler extends CuiHandlerBase implements ICuiComponentHandler {
-    #attribute: string;
-    #args: CuiScrollAttribute;
+export class CuiScrollArgs implements ICuiParsable {
+    target?: string;
+    parent?: string;
+    behavior?: 'auto' | 'smooth';
+
+    constructor() {
+        this.target = null;
+        this.parent = null;
+        this.behavior = 'auto';
+    }
+
+    parse(val: any): void {
+        this.target = getStringOrDefault(val.target, "");
+        this.parent = getStringOrDefault(val.parent, "");
+        this.behavior = is(val.behavior) && val.behavior.toLowerCase() === 'smooth' ? 'smooth' : 'auto';
+    }
+
+
+}
+
+export class CuiScrollHandler extends CuiHandler<CuiScrollArgs> implements ICuiComponentHandler {
     #parent: HTMLElement;
     #target: HTMLElement;
     constructor(element: HTMLElement, utils: CuiUtils, attribute: string) {
-        super("CuiScrollHandler", element, utils);
+        super("CuiScrollHandler", element, new CuiScrollArgs(), utils);
         this.element = element;
-        this.#attribute = attribute
-        this.#args = null;
         this.#parent = null;
         this.#target = null;
     }
 
-    handle(args: any): void {
-        this.parseArguments(args);
+    onInit(): void {
         this.element.addEventListener('click', this.onClick.bind(this));
+        this.setTargets();
     }
-
-    refresh(args: any): void {
-        this.parseArguments(args);
+    onUpdate(): void {
+        this.setTargets();
     }
-
-    destroy(): void {
+    onDestroy(): void {
         this.element.removeEventListener('click', this.onClick.bind(this));
     }
 
@@ -62,7 +76,7 @@ export class CuiScrollHandler extends CuiHandlerBase implements ICuiComponentHan
         let by = to - from;
         this.#parent.scrollBy({
             top: by,
-            behavior: this.#args.behavior ?? 'auto'
+            behavior: this.args.behavior
         });
         this.emitEvent(EVENTS.ON_SCROLL, {
             to: to,
@@ -74,11 +88,10 @@ export class CuiScrollHandler extends CuiHandlerBase implements ICuiComponentHan
         ev.preventDefault();
     }
 
-    private parseArguments(args: any): void {
-        this.#args = args;
-        this.#target = document.querySelector(this.#args.target) as HTMLElement;
+    private setTargets(): void {
+        this.#target = document.querySelector(this.args.target) as HTMLElement;
         if (is(this.#target)) {
-            this.#parent = is(this.#args.parent) ? document.querySelector(this.#args.parent) : this.#target.parentElement;
+            this.#parent = is(this.args.parent) ? document.querySelector(this.args.parent) : this.#target.parentElement;
         }
     }
 }